main.py
import os
from dotenv import load_dotenv
import discord
from discord.ext import commands
import asyncio

load_dotenv()

intents = discord.Intents.all()
bot = commands.Bot(command_prefix='$', intents=intents)

async def load_cogs():
    for filename in os.listdir('./features'):
        if filename.endswith('.py'):
            try:
                await bot.load_extension(f'features.{filename[:-3]}')
                print(f"Loaded {filename}")
            except Exception as e:
                print(f"Failed to load {filename}: {e}")

@bot.event
async def on_ready():
    print(f'Logged in as {bot.user}')

async def main():
    await load_cogs()
    await bot.start(os.getenv('DISCORD_TOKEN'))

if __name__ == "__main__":
    asyncio.run(main())
