requirements.txt
discord.py
python-dotenv
ollama
yt-dlp